requirements.txt
polygon-api-client
plotly
pandas
requests





